refactor(Sim): make sim a package
